Flyttet airport og infostation til domain
